Xenalib: Conversion to run under Python2 and Python3 without flake8
There was need for backwards compatible XenaPythonLib Python3 since many distributions
have been switching to default with Python3.

The main difference between Python2 and Python3 is the basic types that exist to deal
with texts and bytes. On Python3 we have one text type: "str" which holds Unicode data
and two byte types "bytes" and "bytearray".

On the other hand on Python2 we have two text types:"str" which for all intents and
purposes is limited to ASCII + some undefined data above the 7 bit range, unicode
which is equivalent to the Python3 "str" type and one byte type "bytearray" which it
inherited from Python3.

Looking at that you can see that Python3 removed the support for non Unicode data text
and in return gained a hashable byte type, the bytes "object.bytearray".

Unfortunately, Python3 doesn't provide a way to interpret all strings as a particular
type, so the conversion looks a bit ugly. Apologies.

Also much of the conversion here was provided by the "2to3" utility and everything
works fine as of now.

Signed-off-by: Nandan Kulkarni <nakulkar@redhat.com>